enhance: skip return segment/channel info if no changes happens.
querycoord call getRecoveryInfo to get segment/channel list for every
3s, and it will get same content if no changes happens in datacoord,
which will cause a meaningless network and cpu cost.

to solve this, we compute md5 signature for segment list, and use this
signature to check whether segment list changes, if no changes happens,
skip return the segment info in rpc.

Signed-off-by: Wei Liu <wei.liu@zilliz.com>